requirements.txt
aiofiles==23.1.0
aiohappyeyeballs==2.6.1
aiohttp
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
cryptography==44.0.0
defusedxml==0.8.0rc2
dj-database-url==0.5.0
Django==5.1.6
django-ckeditor==6.1.0
django-debug-toolbar==5.0.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-storages==1.14.4
docutils==0.21.2
enum-tools==0.13.0
environ==1.0
environs==5.2.1
Flask==3.1.1
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.4.1
google-cloud-storage==2.7.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
igdb-api-v4==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
ngrok==1.4.0
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.51
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
twitchAPI==4.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
whitenoise==5.2.0
yarl==1.18.3
